Backend/requirements.txt
Django
djangorestframework
django-filter
django-cors-headers
numpy
scikit-learn
pandas
hazm
gensim
dill